Extracts steps
