+++ frontend/src/metabase/query_builder/components/notebook/NotebookStep/NotebookStep.tsx
@@ -17,15 +17,6 @@ import type StructuredQuery from "metabase-lib/queries/StructuredQuery";
 import NotebookStepPreview from "../NotebookStepPreview";
 
 import { NotebookStep as INotebookStep } from "../lib/steps.types";
-import DataStep from "../steps/DataStep";
-import JoinStep from "../steps/JoinStep";
-import ExpressionStep from "../steps/ExpressionStep";
-import FilterStep from "../steps/FilterStep";
-import AggregateStep from "../steps/AggregateStep";
-import BreakoutStep from "../steps/BreakoutStep";
-import SummarizeStep from "../steps/SummarizeStep";
-import SortStep from "../steps/SortStep";
-import LimitStep from "../steps/LimitStep";
 import {
   StepActionsContainer,
   StepBody,
@@ -35,84 +26,7 @@ import {
   StepRoot,
 } from "../NotebookStep.styled";
 
-type StepUIItem = {
-  title: string;
-  icon?: string;
-  priority?: number;
-  transparent?: boolean;
-  compact?: boolean;
-  getColor: () => string;
-
-  // Remove any once all step components are typed
-  component: React.ComponentType<any>;
-};
-
-// TODO
-const STEP_UI: Record<string, StepUIItem> = {
-  data: {
-    title: t`Data`,
-    component: DataStep,
-    getColor: () => c("brand"),
-  },
-  join: {
-    title: t`Join data`,
-    icon: "join_left_outer",
-    component: JoinStep,
-    priority: 1,
-    getColor: () => c("brand"),
-  },
-  expression: {
-    title: t`Custom column`,
-    icon: "add_data",
-    component: ExpressionStep,
-    transparent: true,
-    getColor: () => c("bg-dark"),
-  },
-  filter: {
-    title: t`Filter`,
-    icon: "filter",
-    component: FilterStep,
-    priority: 10,
-    getColor: () => c("filter"),
-  },
-  summarize: {
-    title: t`Summarize`,
-    icon: "sum",
-    component: SummarizeStep,
-    priority: 5,
-    getColor: () => c("summarize"),
-  },
-  aggregate: {
-    title: t`Aggregate`,
-    icon: "sum",
-    component: AggregateStep,
-    priority: 5,
-    getColor: () => c("summarize"),
-  },
-  breakout: {
-    title: t`Breakout`,
-    icon: "segment",
-    component: BreakoutStep,
-    priority: 1,
-    getColor: () => c("accent4"),
-  },
-  sort: {
-    title: t`Sort`,
-    icon: "smartscalar",
-    component: SortStep,
-    compact: true,
-    transparent: true,
-    getColor: () => c("bg-dark"),
-  },
-  limit: {
-    title: t`Row limit`,
-    icon: "list",
-    component: LimitStep,
-    compact: true,
-    transparent: true,
-    getColor: () => c("bg-dark"),
-  },
-};
+import { STEP_UI } from "./steps";
 
 function getTestId(step: INotebookStep) {
   const { type, stageIndex, itemIndex } = step;
